pyproject.toml
[project]
name = "unsolved-mystery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt==3.2.2",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.7",
    "groq>=1.0.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "ollama>=0.6.1",
    "passlib==1.7.4",
    "psycopg2-binary>=2.9.11",
    "pypdf>=6.7.0",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "sentence-transformers==2.7.0",
    "sqlalchemy>=2.0.46",
    "streamlit>=1.54.0",
    "transformers==4.41.2",
    "uvicorn>=0.40.0",
]
